Added an effect that colors a cube using the Geometry Shader which uses less data than the previous CubeSolidScene to produce the same result. Demo scene testing it also added.

+++ Engine/DefaultGeometryShader.h
@@ -8,7 +8,7 @@ class DefaultGeometryShader
 public:
 	typedef Vertex Output;
 public:
-	Triangle<Output> operator()(const Vertex& v0, const Vertex& v1, const Vertex& v2, unsigned int triangle_index) const
+	Triangle<Output> operator()(const Vertex& v0, const Vertex& v1, const Vertex& v2, size_t triangle_index) const
 	{
 		return { v0, v1, v2 };
 	}
